Icon on NftRequirementCard
