+++ src/components/[guild]/Requirements/components/NftRequirementCard/index.ts
@@ -1,0 +1,3 @@
+import NftRequirementCard from "./NftRequirementCard"
+
+export default NftRequirementCard
